Created backing class for PrawnRunFractionParser
Created `PrawnFileRunFractionParser` as a backing class for
`PrawnRunFractionParser`.
